+++ src/imio/events/core/contents/event/content.py
@@ -36,6 +36,7 @@ class EventCroppingProvider(BaseCroppingProvider):
             return [
                 "portrait_affiche",
                 "paysage_affiche",
+                "carre_affiche",
             ]
         return []